# This enables  /Ob3 under Visual Studio 2019