--- a/cmake/simdjson-flags.cmake
+++ b/cmake/simdjson-flags.cmake
@@ -82,7 +82,7 @@ else()
   target_compile_options(simdjson-internal-flags INTERFACE /WX /W3 /sdl)
 endif()
 if(${CMAKE_VS_PLATFORM_TOOLSET} STRGREATER "v141")
-  string(REPLACE "/Ob2" "/Ob2" CMAKE_CXX_FLAGS_RELEASE "${CMAKE_CXX_FLAGS_RELEASE}") #VS 2019
+  string(REPLACE "/Ob2" "/Ob3" CMAKE_CXX_FLAGS_RELEASE "${CMAKE_CXX_FLAGS_RELEASE}") #VS 2019
   message(STATUS " Visual Studio 2019 detected, new compile flags ${CMAKE_CXX_FLAGS_RELEASE}")
 endif()