--- a/cmake/simdjson-flags.cmake
+++ b/cmake/simdjson-flags.cmake
@@ -74,18 +74,17 @@ set(THREADS_PREFER_PTHREAD_FLAG ON)
 
 
 if(MSVC)
-if(${CMAKE_VS_PLATFORM_TOOLSET} STREQUAL "v140")
+if(${MSVC_TOOLSET_VERSION} STREQUAL "140")
   # Visual Studio 2015 issues warnings and we tolerate it,  cmake -G"Visual Studio 14" ..
   target_compile_options(simdjson-internal-flags INTERFACE /W0 /sdl)
 else()
   # Recent version of Visual Studio expected (2017, 2019...). Prior versions are unsupported.
   target_compile_options(simdjson-internal-flags INTERFACE /WX /W3 /sdl)
 endif()
-if(${CMAKE_VS_PLATFORM_TOOLSET} STRGREATER "v141")
+if(${MSVC_TOOLSET_VERSION} STRGREATER "141")
   string(REPLACE "/Ob2" "/Ob3" CMAKE_CXX_FLAGS_RELEASE "${CMAKE_CXX_FLAGS_RELEASE}") #VS 2019
   message(STATUS " Visual Studio 2019 detected, new compile flags ${CMAKE_CXX_FLAGS_RELEASE}")
 endif()
-
 
 else()
   target_compile_options(simdjson-internal-flags INTERFACE -fPIC)