@myd7349 2020-07-04
Hi! @lemire Please consider use `MSVC_TOOLSET_VERSION` instead of `CMAKE_VS_PLATFORM_TOOLSET`. Or quote `CMAKE_VS_PLATFORM_TOOLSET` at least. `CMAKE_VS_PLATFORM_TOOLSET` might be empty if we use `Ninja`: https://github.com/simdjson/simdjson/pull/1009