@lemire 2020-07-04
@myd7349 

> Please consider use MSVC_TOOLSET_VERSION instead of CMAKE_VS_PLATFORM_TOOLSET. Or quote CMAKE_VS_PLATFORM_TOOLSET at least. CMAKE_VS_PLATFORM_TOOLSET might be empty if we use Ninja: #1009

Got it. Fixed.

Why are these variables present in the first place?