@lemire 2020-07-06
I think that this is useless for us. But I am leaving it there nonetheless because Ob3 does have some benefits in other cases.