@lemire 2020-07-09
Dropping.